requirements.txt
Flask==2.3.3
tensorflow-cpu==2.15.0
Pillow==10.1.0
numpy==1.24.3
flask-cors==4.0.0
gunicorn==21.2.0